Better foreground logging (printing).
When run in the foreground, you'll see printed messages.  For non-errors
receiving metrics, you'll still see one line printed per metric received,
but it will include the metric value parts as well.

If there's an error on metric receipt, you'll see that as well (without having
to enable "debug" in the config file.  This is in the spirit of seeing some
output printed regardless of the "debug" value in the config.  Plus, this way
you won't miss StatsD packet decoding errors (which should be rare) if you
ran the daemon with "-f" but didn't have debugging enabled.